Abstraido atraves da classe CreateEntry comandos de criacao
close #10
