+++ src/com/rtjvm/scala/oop/files/DirEntry.scala
@@ -5,5 +5,7 @@ abstract class DirEntry(val parentPath: String, val name: String) {
 
   def asDirectory: Directory
 
+  def asFile: File
+
   def getType: String
 }
