+++ src/com/rtjvm/scala/oop/files/Directory.scala
@@ -1,5 +1,7 @@
 package com.rtjvm.scala.oop.files
 
+import com.rtjvm.scala.oop.filesystem.FileSystemException
+
 import scala.annotation.tailrec
 
 class Directory(override val parentPath: String, override val name: String, val contents: List[DirEntry])
@@ -34,6 +36,7 @@ class Directory(override val parentPath: String, override val name: String, val 
 
   override def getType: String = "Directory"
 
+  override def asFile: File = throw new FileSystemException("A directory cannot be converted to a file!")
 }
 
 object Directory {
