+++ src/com/rtjvm/scala/oop/filesystem/FileSystemException.scala
@@ -1,0 +1,5 @@
+package com.rtjvm.scala.oop.filesystem
+
+class FileSystemException(message: String) extends RuntimeException(message) {
+
+}
